+++ packages/mask/src/extension/popups/pages/Personas/Logout/index.tsx
@@ -1,17 +1,17 @@
 import { memo, useCallback, useMemo, useState } from 'react'
-import { MaskColorVar, makeStyles } from '@masknet/theme'
+import { makeStyles } from '@masknet/theme'
 import { Icons } from '@masknet/icons'
-import { Button, Typography, Box, List, ListItem, Stack, Link } from '@mui/material'
+import { Button, Typography, Box } from '@mui/material'
 import { useI18N } from '../../../../../utils/index.js'
 import { useAsyncFn } from 'react-use'
 import { PersonaContext } from '../hooks/usePersonaContext.js'
 import Services from '../../../../service.js'
 import { LoadingButton } from '@mui/lab'
 import { useNavigate } from 'react-router-dom'
-import { PopupRoutes, type PersonaInformation, NetworkPluginID, formatPersonaFingerprint } from '@masknet/shared-base'
+import { PopupRoutes, type PersonaInformation, NetworkPluginID } from '@masknet/shared-base'
 import { PasswordField } from '../../../components/PasswordField/index.js'
 import { useTitle } from '../../../hook/useTitle.js'
-import { useChainContext, useWallet, useWallets, useWeb3Connection, useWeb3State } from '@masknet/web3-hooks-base'
+import { useWallet, useWallets, useWeb3Connection, useWeb3State } from '@masknet/web3-hooks-base'
 import { isSameAddress, Wallet } from '@masknet/web3-shared-base'
 import { formatEthereumAddress, ProviderType } from '@masknet/web3-shared-evm'
 import { Trans } from 'react-i18next'
@@ -19,7 +19,7 @@ import { first } from 'lodash-es'
 import { useContainer } from 'unstated-next'
 import { PopupContext } from '../../../hook/usePopupContext.js'
 import { WalletRPC } from '../../../../../plugins/Wallet/messages.js'
-import { FormattedAddress } from '@masknet/shared'
+import { ManageWallet } from '@masknet/shared'
 
 const useStyles = makeStyles()((theme) => ({
     content: {
@@ -40,6 +40,7 @@ const useStyles = makeStyles()((theme) => ({
         color: '#FF5555',
         fontSize: 13,
         lineHeight: '18px',
+        marginTop: 16,
     },
     controller: {
         padding: '0 16px 16px 16px',
@@ -159,27 +160,22 @@ export const LogoutUI = memo<LogoutUIProps>(
                     <Icons.CircleWarning size={64} />
                     <Typography className={classes.title}>{t('popups_persona_logout')}</Typography>
                     <Box sx={{ paddingTop: 1.5, paddingBottom: 3, width: '100%' }}>
-                        <ManageWallet manageWallets={manageWallets} />
+                        <ManageWallet manageWallets={manageWallets} persona={selectedPersona} />
                     </Box>
                     <Typography className={classes.tips}>{t('popups_persona_disconnect_tip')}</Typography>
                     {manageWallets.length ? (
-                        <>
-                            <br />
-                            <Typography className={classes.tips}>
-                                <Trans
-                                    i18nKey="popups_persona_disconnect_manage_wallet_warning"
-                                    values={{
-                                        addresses: manageWallets
-                                            .map((x) => formatEthereumAddress(x.address, 4))
-                                            .join(','),
-                                        persona: selectedPersona?.nickname ?? '',
-                                    }}
-                                    components={{
-                                        span: <Typography component="span" sx={{ wordBreak: 'break-word' }} />,
-                                    }}
-                                />
-                            </Typography>
-                        </>
+                        <Typography className={classes.tips}>
+                            <Trans
+                                i18nKey="popups_persona_disconnect_manage_wallet_warning"
+                                values={{
+                                    addresses: manageWallets.map((x) => formatEthereumAddress(x.address, 4)).join(','),
+                                    persona: selectedPersona?.nickname ?? '',
+                                }}
+                                components={{
+                                    span: <Typography component="span" sx={{ wordBreak: 'break-word' }} />,
+                                }}
+                            />
+                        </Typography>
                     ) : null}
                 </div>
 
@@ -233,112 +229,5 @@ export const LogoutUI = memo<LogoutUIProps>(
         )
     },
 )
-
-const useWalletsStyles = makeStyles<{ length: number }>()((theme, props) => ({
-    persona: {
-        padding: '8px 16px',
-        display: 'flex',
-        gap: 20,
-        backgroundColor: MaskColorVar.publicBackground,
-        borderRadius: 8,
-    },
-    nickname: {
-        lineHeight: '16px',
-        fontWeight: 600,
-        fontSize: 14,
-        color: theme.palette.maskColor.primary,
-    },
-    finger: {
-        lineHeight: '16px',
-        fontWeight: 400,
-        fontSize: 12,
-        color: theme.palette.maskColor.primary,
-    },
-    wallets: {
-        gridGap: '12px 12px',
-        display: 'grid',
-        gridTemplateColumns: `repeat(${props.length > 1 ? 2 : 1}, 1fr)`,
-        marginTop: 12,
-        padding: 0,
-    },
-    wallet: {
-        padding: '8px',
-        display: 'flex',
-        gap: 8,
-        backgroundColor: MaskColorVar.publicBackground,
-        borderRadius: 8,
-        flexDirection: 'row',
-        alignItems: 'center',
-    },
-    link: {
-        cursor: 'pointer',
-        zIndex: 1,
-        '&:hover': {
-            textDecoration: 'none',
-        },
-        lineHeight: 0,
-        marginLeft: 6,
-    },
-}))
-interface ManageWalletProps {
-    manageWallets: Wallet[]
-}
-
-function ManageWallet({ manageWallets }: ManageWalletProps) {
-    const { classes } = useWalletsStyles({ length: manageWallets.length })
-    const { selectedPersona } = PersonaContext.useContainer()
-
-    return (
-        <Box>
-            <Box className={classes.persona}>
-                <Box>
-                    <Icons.Masks />
-                </Box>
-                <Stack justifyContent="center">
-                    <Typography variant="body1" className={classes.nickname}>
-                        {selectedPersona?.nickname}
-                    </Typography>
-                    <Typography variant="caption" className={classes.finger}>
-                        {formatPersonaFingerprint(selectedPersona?.identifier.rawPublicKey ?? '')}
-                    </Typography>
-                </Stack>
-            </Box>
-            {manageWallets.length ? (
-                <List className={classes.wallets}>
-                    {manageWallets.map((wallet, i) => (
-                        <WalletItem wallet={wallet} key={i} />
-                    ))}
-                </List>
-            ) : null}
-        </Box>
-    )
-}
-
-interface WalletItemProps {
-    wallet: Wallet
-}
-function WalletItem({ wallet }: WalletItemProps) {
-    const { classes } = useWalletsStyles({ length: 1 })
-    const { Others } = useWeb3State()
-    const { chainId } = useChainContext()
-    return (
-        <ListItem className={classes.wallet}>
-            <Icons.SmartPay />
-            <Stack flexDirection="column">
-                <Typography className={classes.nickname}>{wallet.name}</Typography>
-                <Typography className={classes.finger}>
-                    <FormattedAddress address={wallet.address} size={4} formatter={Others?.formatAddress} />
-                    <Link
-                        className={classes.link}
-                        href={Others?.explorerResolver.addressLink(chainId, wallet.address)}
-                        target="_blank"
-                        rel="noopener noreferrer">
-                        <Icons.LinkOut size={12} sx={{ transform: 'translate(0px, 2px)' }} />
-                    </Link>
-                </Typography>
-            </Stack>
-        </ListItem>
-    )
-}
 
 export default Logout
