+++ packages/shared/src/UI/components/index.ts
@@ -71,3 +71,4 @@ export * from './TokenTransactionConfirmModal/index.js'
 export * from './TokenMenuList/index.js'
 export * from './TokenWithSocialGroupMenu/index.js'
 export * from './NFTFallbackImage/index.js'
+export * from './PersonaWallet/index.js'
